Update patch set 3
Patch Set 3:

Because easy is no fun :D.

In fact I wanted to find a way to do it, but apparently not meant to be done this way.
Even the date field, the only one on normal Android to have a space, is written by a shell script & not an override.

So yeah, back to basics.
Doesn't annoy me if it's there only in AICP, makes the syncs easy.

Patch-set: 3